+++ src/utils/createDataLoader.js
@@ -1,23 +1,7 @@
 const fs = require("fs");
 const path = require("path");
 const combinedData = require("../_data/combineAffiliations");
-const variables = require("./variables");
-
-// const nat = require("../_data/affiliations/nat");
-// const ak = require("../_data/affiliations/ak");
-// const al = require("../_data/affiliations/al");
-// const ar = require("../_data/affiliations/ar");
-// const az = require("../_data/affiliations/az");
-// const dc = require("../_data/affiliations/dc");
-
-// const combinedData = {
-// 	ak,
-// 	al,
-// 	ar,
-// 	az,
-// 	dc,
-// 	nat,
-// };
+const variables = require("./outputVariables");
 
 const minifiedJsonString = "{data = " + JSON.stringify(combinedData) + "}";
 
@@ -28,9 +12,11 @@ const minifiedDistPath = path.join(distDir, "dataLoader.min.json");
 fs.mkdirSync(distDir, { recursive: true });
 
 let data = (minifiedJsonString + variables).replace(/\s/g, "");
+// console.log(data);
+
 fs.writeFile(minifiedDistPath, minifiedJsonString + variables, (err) => {
 	if (err) throw err;
 	console.log("Minified data written to file");
 });
 
-// console.log(data);
+
